Seperate an reusable json-rpc client crate
