+++ client/cli/Cargo.toml
@@ -34,7 +34,7 @@ crash-handler = { path = "../../common/crash-handler", version = "0.1.0" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
 lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-wallet = { path = "../libra-wallet", version = "0.1.0" }
-libra-json-rpc = { path = "../../json-rpc", version = "0.1.0" }
+libra-json-rpc-client = { path = "../json-rpc", version = "0.1.0" }
 libra-logger =  { path = "../../common/logger", version = "0.1.0" }
 libra-metrics = { path = "../../common/metrics", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
