+++ client/json-rpc/src/lib.rs
@@ -1,0 +1,14 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+mod blocking;
+mod client;
+
+pub mod errors;
+pub mod views;
+
+pub use blocking::JsonRpcClient;
+pub use client::{
+    get_response_from_batch, process_batch_response, JsonRpcAsyncClient, JsonRpcBatch,
+    JsonRpcResponse,
+};
