+++ json-rpc/Cargo.toml
@@ -11,17 +11,20 @@ edition = "2018"
 
 [dependencies]
 anyhow = "1.0"
+futures = "0.3.0"
 hex = "0.4.2"
+once_cell = "1.3.1"
+prometheus = { version = "0.8.0", default-features = false }
+reqwest = { version = "0.10.4", features = ["blocking", "json"], default_features = false }
 serde_json = "1.0.51"
 serde = { version = "1.0.106", default-features = false }
-warp = "0.2.2"
-futures = "0.3.0"
-reqwest = { version = "0.10.4", features = ["blocking", "json"], default_features = false }
 tokio = { version = "0.2.13", features = ["full"] }
-once_cell = "1.3.1"
+warp = "0.2.2"
 
-proptest = { version = "0.9.6", optional = true }
+client = { path = "../client/json-rpc", version = "0.1.0", package = "libra-json-rpc-client" }
+debug-interface = { path = "../common/debug-interface", version = "0.1.0" }
 lcs = { path = "../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
+libradb = { path = "../storage/libradb", version = "0.1.0" }
 libra-config = { path = "../config", version = "0.1.0" }
 libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
 libra-logger = { path = "../common/logger", version = "0.1.0" }
@@ -29,13 +32,10 @@ libra-mempool = { path = "../mempool", version = "0.1.0" }
 libra-metrics = { path = "../common/metrics", version = "0.1.0" }
 libra-proptest-helpers = { path = "../common/proptest-helpers", optional = true }
 libra-types = { path = "../types", version = "0.1.0" }
-libra-workspace-hack = { path = "../common/workspace-hack", version = "0.1.0" }
-libradb = { path = "../storage/libradb", version = "0.1.0" }
-move-core-types = { path = "../language/move-core/types", version = "0.1.0" }
-storage-interface = { path = "../storage/storage-interface", version = "0.1.0" }
-transaction-builder = { path = "../language/transaction-builder", version = "0.1.0"}
-debug-interface = { path = "../common/debug-interface", version = "0.1.0" }
 libra-temppath = { path = "../common/temppath", version = "0.1.0", optional = true }
+libra-workspace-hack = { path = "../common/workspace-hack", version = "0.1.0" }
+proptest = { version = "0.9.6", optional = true }
+storage-interface = { path = "../storage/storage-interface", version = "0.1.0" }
 
 [dev-dependencies]
 vm-validator = { path = "../vm-validator", version = "0.1.0" }
